+++ CHANGELOG.rst
@@ -1,4 +1,12 @@
-**0.4.0 - 04/11/23**
+**0.5.0 - 04/13/23**
+ - Bugfix to apply incorrect selection noising at the expected probabilility
+ - Implement the omission noise function
+ - Implement schema for output columns and their dtypes
+ - Implement a year filter to the form generation functions
+ - Support a path to data root directory as form generation function argument
+ - Update documentation
+ 
+ **0.4.0 - 04/11/23**
  - Generate default configuration instead of maintaining a static file
  - Read sample data if no data argument is provided
  - Update sample datasets